+++ lib/wnm_support/active_record_ext/view_helpers.rb
@@ -2,7 +2,7 @@ module WnmSupport
   module ActiveRecordExt
     module ViewHelpers
       def helpers
-        #ActionController::Base.helpers
+        ActionController::Base.helpers
       end
     end
   end