+++ lib/wnm_support/view_helpers/bool_to_human.rb
@@ -1,0 +1,15 @@
+module WnmSupport
+  module ViewHelpers
+    module BoolToHuman
+      def bool_to_human(value)
+        I18n.t(value == true ? "yes" : "no")
+      end
+
+      def bool_to_human_with_empty(value)
+        return "" if value.nil?
+
+        bool_to_human(value)
+      end
+    end
+  end
+end