+++ spec/wnm_support/activer_record_ext/mysql_order_by_field_spec.rb
@@ -1,5 +1,7 @@
 require 'spec_helper'
 
+ActiveRecord::Base.send :include, WnmSupport::ActiveRecordExt::MysqlOrderByField
+
 class News < ActiveRecord::Base
 
 end