+++ spec/wnm_support/activer_record_ext/view_helpers_spec.rb
@@ -1,5 +1,7 @@
 require "spec_helper"
 
+ActiveRecord::Base.send :include, WnmSupport::ActiveRecordExt::ViewHelpers
+
 class Item < ActiveRecord::Base
   include ActiveRecordNoTable
 end